Adds BundleAppId to ReactInstanceSettings
For scenarios where you may want to run a single packager server
instance to serve multiple apps with different rules, and the relative
bundle path is not sufficient for determining which application the
bundle is being produced for, Metro provides an `app` query parameter to
the dev server.

This change wires up the option for apps to supply the `app` query
parameter to metro via ReactInstanceSettings::BundleAppId. This
functionality is already available to iOS and Android React Native apps.