Avoid copying ReactPromise in ImageLoader module
We're seeing crashes when the React instance is destroyed if there is an active ImageLoader::getSizeWithHeaders request in flight (related to JSI trying to create an object on a destroyed runtime).

This is similar to an issue that was resolved and picked into React Native Windows v0.68 (#10436), but even with that fix, the crash still occurs. I suspect it may be related to using the ReactPromise copy constructor, but even if not, it's not so bad to avoid copying the ReactPromise.