[0.70] CrashManager: handle abort signal
This PR backports #11374 to 0.70.

If there is a noexcept function on the stack, the CRT just calls abort instead of calling the handler registered via SetUnhandledExceptionFilter.

This change registers a handler for the SIG_ABRT (the abort signal) to attempt to collect crash details.

As before, nothing happens by default, this code path is only enabled if EnableDefaultCrashHandler is set to true.