Workaround fix for "missing-asset-registry-path" error to unblock PRs
New app projects built against main are unable to create their bundle (see #11437).

As such, all PRs are blocked as the New CLI app checks all fail.

The problem is that the metro config needs `transformer.assetRegistryPath` to be set to the location of `./Libraries/Image/AssetRegistry`, which for our purposes is the copy located in the RNW folder. For some reason this isn't set anymore.

This PR adds a workaround of specifying that path in our app template's metro.config.js file. This unblocks our PRs but we should probably find a better place than in an app template file.

Resolves #11437