Ensure ReactNativeWindowsDir is set in Directory.Build.props
This PR fixes a problem where we use the `ReactNativeWindowsDir` prop to
determine where to build intermediate and output files, but we do so
*before* verifying that the prop has been set.

This regression was introduced in RNW 0.68 and the result is that rather
than all build outputs being placed under the root `build` and `target`
folders, they are placed deeper in the folder structure. This can cause
file paths to exceed the default max path length, which can cause build
problems as not every tool supports long paths (even if it's enabled in
the system).