Use `--sourcemap-use-absolute-path` when creating release bundles
The last string at the bottom of a bundle is the path to the sourcemap
for that bundle. Right now, we just put the name of the sourcemap,
assuming that it will be kept side-by-side with the bundle. However by
default we don't put the sourcemap side-by-side with the bundle output.

This change to use `--sourcemap-use-absolute-path` will put an absolute
path to the sourcemap, which makes it possible to direct debug bundles
when using Chakra, with Visual Studio.

Note: Hermes direct debug always requires Metro, so this change doesn't
improve the Hermes debugging experience.

Also, this shouldn't affect many internal CI systems as they often
overwrite this line anyway as sourcemaps get moved to shared build
servers.

Closes #11235