Fix typos and add `cspell` configuration
This PR adds configuration files for `cspell`, which is available as a recommended VS Code plugin for this repo. This PR also fixes most "non-breaking" typos (comments, error messages, local names) for code files within the repo.

This PR **does not** add any kind of spell-checking enforcement on the repo.