Improve Chocolatey detection in rnw-dependencies.ps1
This PR improves the `rnw-dependencies.ps1` script to do a better job
detecting Chocolately's choco command.

Specifically, if the command is not in the `PATH` for some reason, the
script will now try to look in the default install location (as defined
by the official chocolately install script).

This PR also makes sure that we re-check for choco.exe after running its
install script, and I've also refactored direct calls to choco.exe into
a function that ensures that chocolately is available first.

I've tested installs in each scenario:

* Not having chocolately installed
* Having chocolately installed with choco.exe in the `PATH`
* Having chocolately installed without choco.exer in the `PATH`

Closes #11207