Fix typos and add `cspell` configuration  (#11536)
## Description

This PR adds configuration files for `cspell`, which is available as a recommended VS Code plugin for this repo. This PR also fixes most "non-breaking" typos (comments, error messages, local names) for code files within the repo.

This PR **does not** add any kind of spell-checking enforcement on the repo.

### Type of Change
N/A

### Why

Clear up misunderstandings in comments / error messages.

### What
What changes were made to the codebase to solve the bug, add the functionality, etc. that you specified above.

## Screenshots
N/A

## Testing
N/A