[0.72] Fix integration test failures in Debug
This PR backports #11596 to 0.72.

This PR removes the `Slider` test case in `MountComponentTests.tsx`.

This test has been marked to skip for almost three years now (so I
imagine it never worked).

Even though the test was never run, recently just the act of trying to
load the `Slider` component within the JS causes an uncaught exception
to be thrown in debug builds, which blocks the whole test suite from
running.

Closes #11556
Closes #11562