Fix a bunch of properties on Text (#11605)
At some point we lost a bunch of properties on Text -- including accessibilityLabel. This should light up those properties again.

During various RN merges - we brought in changes that extracted more props out of "restProps", but didn't merge in addition property sets on the components, so we lost the values of a bunch of properties.