Fix dependency script to install missing components for VS Preview
The dependency script could detect that a Preview install meets the
version requirement, but when it came to installing any missing VS
components, would fail to remember to use the Preview install.

Best case it would try to install a stable version and add the
components to that instead.

This does not solve the problem of not being able to install Preview
builds with Choco.

Closes #11665