Limit Yoga layout to relevant root on resize (#10580)
* Limit Yoga layout to relevant root on resize

Recalculation of Yoga layout due to root view size change does not
require that we re-run layout on all roots. It also does not require
that we invoke UpdateLayout on any new nodes (as a resize operation is
independent of a batch of UIManager operations).

This change limits the scope of Yoga calculation and application of Yoga
results to XAML views to only the root view that changed size. This
optimization primarily impacts apps with multiple windows, but can also
benefit any React Native Windows app with multiple roots.

* Change files