[0.70] Disable RestoreUseStaticGraphEvaluation for VS >= 17.6
This PR backports #11681 to 0.70.

This PR disables `RestoreUseStaticGraphEvaluation` if building within Visual Studio >= 17.6, which is not working in our project.

Ideally we want `RestoreUseStaticGraphEvaluation` to improve the build speed. However this will unblock folk trying to build our code within the newest versions of Visual Studio.

This does not impact building via MSBuild (i.e. using the `run-windows` CLI command).

Resolves #11670

- Bug fix (non-breaking change which fixes an issue)

Unblock building with VS >= 17.6

Resolves #11670

Adds a check for the version of VS to the condition setting `RestoreUseStaticGraphEvaluation`.

N/A

Verified I can build `Microsoft.ReactNative.sln` with VS 17.6 and 17.5.