[0.72] Migrate eslint-config from `@react-native-community` to `@react-native` (#11746)
\## Description

This PR upgrades the dependency as it was re-named upstream.

\### Type of Change
- Bug fix (non-breaking change which fixes an issue)

\### Why
What is the motivation for this change? Add a few sentences describing
the context and overall goals of the pull request's commits.

Closes #11742

\### What
Changed the dependencies in `@rnw-scripts/eslint-config`'s package.json.

\## Screenshots
N/A

\## Testing
Ran `yarn lint` and saw no changes.