Adds ABIViewManager interface for 3p popup content
We discovered a bug where a custom component that hoists it's children
into a popup has issues with Pressable. Specifically, Pressable calls
UIManager.measure to test whether a pointer is still in the hit box for
the first responder after pointerDown on each pointerMove.

FlyoutViewManager works around this by returning `true` from
`ShadowNodeBase::IsWindowed()`. There is no affordance for this in
ABIViewManager. This change introduces an interface to support it.