Fix accessibilityState to support states not being set (#11792)
## Description

Our existing implementation for `accessibilityState` does not support that the values of a state can be `null`. This means *every* View has *every* unset state with the default `false` value.

The problem stems from implementing the short-lived `accessibilityStates` prop at the same time as `accessibilityState`, but not removing the deprecated API right away.

This PR fixes our implementation to only include the states when explicitly set, and to otherwise clear them. This PR also removed the no longer required "opposite" states like `Collapsed` and `Unchecked`.

Resolves #11791

### Type of Change
- Breaking change (fix or feature that would cause existing functionality to not work as expected)

### Why
Properly implement accessibility in preparation for doing the work correctly for Fabric.

Resolves #11791

### What

Removed the old unused states and updated FrameworkElementViewManager to allow for unset values rather than forcing the default false. Updated AccessbilityStateChecked to accept an enum of Unchecked, Mixed, or Checked. Updated the DynamicAutomationPeer to only express a View has the appropriate provider iff the matching state is present on the view.

## Screenshots
N/A

## Testing

Updated E2E test snapshots to include the AccessibilityRole and AccessibilityState properties in the XAML tree dumps.