Deprecate Microsoft.ReactNative.WindowsAppSDK project (#11945)
## Description

This PR deprecates the `Microsoft.ReactNative.WindowsAppSDK` project.

### Type of Change
- Breaking change (fix or feature that would cause existing functionality to not work as expected)
- This change requires a documentation update

### Why

These projects/templates were early, experimental work for a future of RNW apps being WinAppSDK-based. Since we're not going in that direction, we shouldn't persist allowing projects to be created using it.

Resolves #11744

### What

* Deleting the `Microsoft.ReactNative.WindowsAppSDK` project files
* Removing all jobs/tasks related to building/testing/publishing `Microsoft.ReactNative.WindowsAppSDK`
* Removes the `cs-app-WinAppSDK` app template and specific new project creation code

**Note:** The existing `UseWinUI3` msbuild property and corresponding `USE_WINUI3` build flags are intentionally not touched in this PR, to maintain the minimum support for some customers.

## Screenshots
N/A

## Testing
N/A

## Changelog
Should this change be included in the release notes: Yes

The Microsoft.ReactNative.WindowsAppSDK experimental project has been deprecated and will no longer be supported.