[0.72] Properly detect the correct machine architecture in CLI
This PR backports #11978 to RNW 0.72.

## Description

Node's `process.arch` and `os.arch()` methods both return the architecture that the node binaries were built for, not the architecture of the machine they're running.

This can cause issues with the various hardware virtualization layers, specifically when running 32-bit Node on 64-bit machines.

Closes: #11977

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
We keep hitting issues when users run the 32-bit Node on 64-bit Windows configuration.

Closes: #11977

### What

This PR:
* Updates and exports the `deviceArchitecture()` API in the `@react-native-windows/telemetry` library, which now correctly detects when x86 Node is being used on an x64 machine
* Adds and exports a new `nodeArchitecture()` method, which is also now recorded in telemetry
* Updates the `react-native-windows-init` and `@react-native-windows/cli` commands to warn the user if there's a mismatch
* Fixes an issue where querying for a registry key would fail due to querying the wrong registry because of the SysWOW64 redirection

## Screenshots
N/A

## Testing

Updated the telemetry tests with a new test for `nodeArchitecture`.

## Changelog
Should this change be included in the release notes: Yes

Properly detect the correct machine architecture for 32-bit Node on 64-bit Windows scenarios