[0.71] Fix BinSkim warnings for Desktop
This PR backports #11816 to 0.71.

BinSkim has been throwing warnings for Desktop ever since the upgrade to BinSkim@4. Particularly this is from scanning the Hermes and V8 dlls.

However both are sourced from their own Microsoft-owned repos and only consumed here via published NuGets. BinSkim should be running on those repos (if required) not here in RNW.

This PR updates the pipeline to only test the binaries built by this pipeline.