Fix bug when blur command received on node without XamlRoot
We are seeing crashes on quit when the blur command is called roughly
around the same time that the React instance is being shut down. It's
likely because the result of tryGetXamlRoot is null.

This change falls back on the GetFocusedElement() overload without any
parameters if the XamlRoot is null, as supplying a null parameter will
result in a NPE.