Adds support for experimental_layoutConformance to Paper UIManager
This change backports the functionality to enable the
`experimental_layoutConformance` prop to Windows Paper UIManager.

This change does a few things:
1. Creates a per YGNode YGConfig and updates the unique_ptr deleter to
   free the YGConfig when the YGNode is freed.
2. Adds the algorithmic support for handling
   experimental_layoutConformance correctly in Paper.

This change will come "for free" with Fabric, but it might be useful to
evaluate Web-conformant layout in Yoga on Paper.

Adding a per YGNode YGConfig is likely to increase the memory profile of
react-native-windows by 34 bytes per ShadowNode, which shouldn't be a
huge issue as most apps only have a few thousand nodes.

The algorithm for handling experimental_layoutConformance is as follows:
1. When a node gets an explicit prop value for
   experimental_layoutConformance, recursively apply that value to
   current children
2. When a node gets an undefined value for
   experimental_layoutConformance, you can assume that the node
   previously had a defined value. Walk the ancestor tree for the next
   defined experimental_layoutConformance prop and apply recursively to
   all children of that node.
3. When a node is added to another node, check if the parent node has
   applied a layout conformance value, and if so, apply the current
   value to all children.
4. In all recursive cases above, bail out of recursion when you reach a
   node with a defined experimental_layoutConformance value.

The hope of this change is that it only impacts the memory profile of
Paper UIManager, with some additional CPU time spent on allocations and
lookups. But for the most part, so long as
experimental_layoutConformance is never used, the cost of recursing
through the shadow tree will never apply.