Prevent Desktop.IntegrationTests from hanging forever
While trying to integrate the latest RN code (see PR #11970) the `RNTesterIntegrationTests` in `React.Windows.Desktop.IntegrationTests` kept hanging and timing out.

The root issue (TBD) was a `JSError` being thrown by Chakra when the test bundle (ex. `DummyTest`) was loaded from Metro. However the bundle was loaded asynchronously, which caused that error to be swallowed and never reported. Furthermore, not knowing that the bundle load had failed, the test continued onward and attempted to start loading the app with `AppRegistry.runApplication()` (which also failed silently) and finally proceeded to wait indefinitely for the JS component to render and fire the `TestModule.markTestPassed(true)`.

To address this, this PR changes the tests to:
1. Load the bundle synchronously, so if there's a failure it gets thrown and caught by the test runner immediately.
2. Set a reasonable 30s timeout for waiting for the app to "render" (it usually completes in 2-3s).