Prevent Desktop.IntegrationTests from hanging forever (#12072)
## Description

While trying to integrate the latest RN code (see PR #11970) the `RNTesterIntegrationTests` in `React.Windows.Desktop.IntegrationTests` kept hanging and timing out.

The root issue (TBD) was a `JSError` being thrown by Chakra when the test bundle (ex. `DummyTest`) was loaded from Metro. However the bundle was loaded asynchronously, which caused that error to be swallowed and never reported. Furthermore, not knowing that the bundle load had failed, the test continued onward and attempted to start loading the app with `AppRegistry.runApplication()` (which also failed silently) and finally proceeded to wait indefinitely for the JS component to render and fire the `TestModule.markTestPassed(true)`.

To address this, this PR changes the tests to set a reasonable 30s timeout for waiting for the app to "render" (it usually completes in 2-3s).

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why

To better detect failures in `RNTesterIntegrationTests`.

### What

Changes the tests to set a reasonable 30s timeout for waiting for the app to "render" (it usually completes in 2-3s).

## Screenshots
N/A

## Testing
Re-ran the tests and verified if the bundle load fails the test fails.

## Changelog
Should this change be included in the release notes: No