[0.72] Update Policheck exclusions and exceptions 9/15
This PR backports #12139 to 0.72.

It seems the Policheck tasks no longer accept blanket "exclusions" in our rules database for terms we use everywhere. So I've gone and updated the individual "file exceptions" for those terms in our database.

It's harder to maintain, but this should unblock PRs.

- Bug fix (non-breaking change which fixes an issue)

Compliance.

Updated rules database.

N/A

N/A

Should this change be included in the release notes: no