Expose DisableXamlGuard property to allow mixing of WinUI 2/3
You can't mix WinUI 2 and WinUI 3 in the same process, so just to be
safe, RNW implments a guard (see PR #5290) which will cause an explicit
crash if you try, in case an end developer does so untintentionally.

This PR adds an msbuild prop `DisableXamlGuard` which, when set to true,
will remove this guard in the RNW code.