Refactor `init-windows` CI jobs (#12213)
Backport this into 0.73 to keep the pipelines aligned.

PR #12183 show-horned a new variable and branching path in our existing `cli-init.yml` CI job to support the new `react-native init-windows` command.

However, this makes the pipeline hard to parse and maintain. This PR factors out the new new project path into a separate set of jobs, so we can leave the "old" pipeline as-is, run both for now, but have and easier time updating the new one (and eventually just delete the old one when it's no longer necessary).

- New feature (non-breaking change which adds functionality)

To make the new new project CI easier to maintain and drift from the old new project CI.

Created a new `cli-init-windows.yml` job, which calls a new "template-focused" `react-native-init-windows.yml` template, and also factored out code common to both it and to `react-native-init.yml` (like starting/stopping verdaccio) into reusable templates.

N/A

Ran the new pipelines

Should this change be included in the release notes: _no_