[C++ 20 Bump] Fix build artifact uploads for Fabric E2E
A revision of the all-in-one C++ 20 PR caused a crash n Fabric E2E tests. I wanted to look at the crash dump, alongside the pdb and DLL used by the build. The previous copy of this logic, reused for Fabric, was dependent on the structure for a C# appx build. This updates the direcotry based on what I locally observed for the Fabric E2E test build.