C++ 20: Treat `char8_t` as `char` for unicode literals in source
`u8` is needed to portably ensure source is parsed as UTF-8 (though RNW already has build settings to do this)

In C++ 20, these are represented using `char8_t` instead of `char`, which makes them incompatible with the other UTF-8 string APIs.

This wraps C string literals with `reinterpret_cast`, to keep the portablility benefits of `u8` but also store into a `std::string`.