Refactor react/renderer/graphics/Color handling
This moves to the cross-platform abstractions used by mobile. It
requires one minor change Color.cpp, which we should be able to upstream
easily (removing the static const UndefinedColor in SharedColor).

It also has a minor change to the resolution of certain extension APIs
to SharedColor. For example, `AsColorRefWithAlpha` is now a method on
facebook::react::Color instead of facebook::react::SharedColor.

Towards #7821