Fabric: Add `windows` script to new cpp-app template
Adds the functionality for apps created with the new cpp-app template to run `yarn windows`, as an alias for `react-native run-windows`, similar to the old template.

To do this, I've created a new `templateUtils` module for shared template config code, with a new `updateProjectPackageJson()` method to make it easy for templates to inject items into a project's package.json file (in this case, a new `scripts` entry).