Beachball Check should actually stop publish (#12523)
## Description

This PR fixes the beachball check added in #12293 to actually stop the publish pipeline when errors are detected. Furthermore it adds a parameter to enable this check, default on, but which can be manually disabled when we're trying to promote a release branch (either from latest to legacy, preview to latest, etc).

This PR also adds a version of the check to the standard setup task that instead reports warnings when invalid change files are detected - hopefully this will improve our chances of catching these errors before getting to publish, without having to unduly block CI/PRs.

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
In publishing 0.73 an (expected) error was detected as the `@office-iss\react-native-win32` package was being promoted. Any error should have caused the publish to bail early - but also highlighted that when we **want** to promote, we'll need a way to turn the check off.

### What
See above.

## Screenshots
N/A

## Testing
Verified by adding bad change files to this PR (now removed) to see that the check happens.

## Changelog
Should this change be included in the release notes: no