[0.73] Lock to applicationinsights@2.7.3 to prevent using newer broken versions
## Description

When updating dependencies in main, (see #12575), we found that
applicationinsights@2.9.2 breaks our CLI commands.

We aren't actually collecting telemetry so there's no problem with
locking the version of this package in this branch for now.

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
To unblock calls to CLI commands by new projects.

### What
Lock to the current version used by this branch.

## Screenshots
N/A

## Testing
Ran yarn build locally, which calls `codegen-windows` internally.

## Changelog
Should this change be included in the release notes: no