Fix license header in file overrides (#12596)
* Fix license header in file overrides

The upstream license headers switched from:
```
Copyright (c) Facebook, Inc. and its affiliates.
```
To:
```
Copyright (c) Meta Platforms, Inc. and affiliates.
```

This updates all the file overrides with the correct license header.

* Change files

* Adds license header to WindowTextInputNativeComponent.js

Just following the pattern for other files in this folder to add a license header. Likely an oversight that it wasn't added in the first place.