Return null for missing view manager constants
react-native-windows currently returns an empty object for UIManager::getConstantsForViewManager. This breaks the hasViewManagerConfig checks in JS, which may be used to determine whether a view should attempt to render or not (or to gate a requireNativeComponent call, in the case of DebuggingOverlay). Having this behavior would have prevented the need to introduce the dummy DebuggingOverlay view manager.