Fix issue with new arch cpp-app apps not finding the bundle in release builds (#12754)
## Description

The new cpp-app template (and related apps `playground-composition` and `e2e-test-app-fabric`) were all set with an invalid `BundleRootPath`. This causes the app to throw an error because it can't parse the given path from the URI, furthermore, since it was a URI based on the current working directory, it might not even be the right directory to load the bundle from.

This PR fixes the issue by instead using the correct `file://` URI scheme and also looking for the bundle path relative to the location of the app exe itself. This PR also re-aligns the apps to use the similar instance settings setup of the older UWP app (using both the `BUNDLE` and `_DEBUG` macros appropriately).

Partially resolves #12752

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
So that release builds work for the apps using the new template.


### What
See above.

## Screenshots
![image](https://github.com/microsoft/react-native-windows/assets/10852185/4a839f09-996b-464d-ada3-1878bdf835e6)

## Testing
Tested a new app works with the changes.

## Changelog
Should this change be included in the release notes: _yes_

Fixed issue with new arch cpp-app apps not finding the bundle in release builds