Fix `run-windows` to use 64-bit MSBuild (like VS 2022)
## Description

VS 2022 uses the 64-bit MSBuild by default. It's also what we use in CI when building. For some reason, `run-windows` is still using 32-bit MSBuild, which is unnecessary and puts memory limits on our compilation tasks.

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why

In PR #12798 I'm trying to update our version of CppWinRT, and in some scenarios (namely release fabric builds) we've hit the 32-bit memory limit for linking.

### What
Changed the CLI to use the 64-bit msbuild exe instead of the 32-bit one it was hardcoded to use.

## Screenshots
N/A

## Testing
Verified everything still builds.

## Changelog
Should this change be included in the release notes: yes

The CLI build command `run-windows` now uses the same 64-bit MSBuild VS 2022 uses by default