Remove previous SizeChanged event listener when reloading the app (#12888)
* Remove previous SizeChanged event listener when reloading the app

After a dev reload, the ReactRootView is assigned a new root tag. We never evicted the old auto revoker from the m_sizeChangedVector when this occurred, and could lead to attempts to do layout on a stale root tag. This fixes the issue, moving the SizeChanged event revoker data structure to a map, and evicting the listener for the root tag when it is removed.

* Change files