[0.74] Add builds and nugets for MS.RN Fabric
This PR backports #12900 to 0.74.

## Description

This PR adds configurations for Microsoft.ReactNative with Fabric enabled to PR, CI, and Publish builds. This change will enable the publishing of new versions of the Microsoft.ReactNative NuGet with the `-Fabric` suffix to differentiate it from the current Paper NuGet.

This replicates the setup that Desktop has, where we create and publish versions of React Native Win32 nugets with the `-Fabric` suffix to indicate those versions are using Fabric.

### Type of Change
- New feature (non-breaking change which adds functionality)

### Why
We want the default new arch, new app / lib experience to *not* build RNW, but instead consume it via NuGets, see #12639. In order to get that started, we'll need the

### What
Updated CI/PR/Publish configurations.

## Screenshots
N/A

## Testing
Verified new configurations fired off existing tests in PR.

## Changelog
Should this change be included in the release notes: yes

New Microsoft.ReactNative nugets with Fabric enabled