[0.74] Add builds and nugets for MS.RN Fabric and Fixed Microsoft.ReactNative.IntegrationTests and Mso.UnitTests to run with Fabric (#12919)
This PR backports #12900 and #12924 to 0.74.

From #12900:

## Description

This PR adds configurations for Microsoft.ReactNative with Fabric enabled to PR, CI, and Publish builds. This change will enable the publishing of new versions of the Microsoft.ReactNative NuGet with the `-Fabric` suffix to differentiate it from the current Paper NuGet.

This replicates the setup that Desktop has, where we create and publish versions of React Native Win32 nugets with the `-Fabric` suffix to indicate those versions are using Fabric.

### Type of Change
- New feature (non-breaking change which adds functionality)

### Why
We want the default new arch, new app / lib experience to *not* build RNW, but instead consume it via NuGets, see #12639. In order to get that started, we'll need NuGets built with fabric enabled.

### What
Updated CI/PR/Publish configurations.

## Screenshots
N/A

## Testing
Verified new configurations fired off existing tests in PR.

## Changelog
Should this change be included in the release notes: yes

[0.74] New Microsoft.ReactNative nugets with Fabric enabled

From #12924:

## Description

This PR makes Microsoft.ReactNative.IntegrationTests aware of Fabric requirements to use the WinAppSDK `DispatchQueueController`. This PR also fixes both that project and Mso.UnitTests to be able to consume self-contained WinAppSDK so they can run (the tests had code but were technically never exercising it).

Finally this PR also enables a fabric config in PR to make sure we catch test failures like this in the future.

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
See above.

Resolves #12923 

### What
See above.

## Screenshots
N/A

## Testing
Ran Microsoft.ReactNative.Integration tests locally.

## Changelog
Should this change be included in the release notes: _yes_

[0.74] Fixed Microsoft.ReactNative.IntegrationTests and Mso.UnitTests to run with Fabric